+++ java/samples/sample-code/src/main/java/com/microsoft/semantickernel/syntaxexamples/Example08_RetryHandler.java
@@ -14,7 +14,7 @@ import com.microsoft.semantickernel.textcompletion.CompletionSKContext;
 import com.microsoft.semantickernel.textcompletion.CompletionSKFunction;
 import com.microsoft.semantickernel.textcompletion.TextCompletion;
 import com.microsoft.semantickernel.util.AzureOpenAISettings;
-import com.microsoft.semantickernel.util.ClientSettings;
+import com.microsoft.semantickernel.util.AIProviderSettings;
 
 import java.io.IOException;
 import java.time.Duration;
@@ -31,7 +31,7 @@ public class Example08_RetryHandler {
                 .setMaxRetries(3)
         );
 
-        AzureOpenAISettings settings = ClientSettings.getAzureOpenAISettingsFromFile(CONF_PROPERTIES);
+        AzureOpenAISettings settings = AIProviderSettings.getAzureOpenAISettingsFromFile(CONF_PROPERTIES);
         AzureOpenAIClient client = new AzureOpenAIClient(
                 new OpenAIClientBuilder()
                         .retryOptions(retryOptions)
